+++ app/controllers/facilities_management/rm6378/admin/users_controller.rb
@@ -1,0 +1,9 @@
+module FacilitiesManagement
+  module RM6378
+    module Admin
+      class UsersController < Base::UsersController
+        include FacilitiesManagement::Admin::FrameworkStatusConcern
+      end
+    end
+  end
+end