+++ app/controllers/facilities_management/rm6378/sessions_controller.rb
@@ -1,0 +1,6 @@
+module FacilitiesManagement
+  module RM6378
+    class SessionsController < FacilitiesManagement::SessionsController
+    end
+  end
+end